Rename HistoricalPrice to Price
