+++ src/main/java/com/jespinel/stockreader/clients/StockAPIClient.java
@@ -1,6 +1,6 @@
 package com.jespinel.stockreader.clients;
 
-import com.jespinel.stockreader.entities.HistoricalPrice;
+import com.jespinel.stockreader.entities.Price;
 import com.jespinel.stockreader.entities.Stats;
 import com.jespinel.stockreader.entities.Symbol;
 
@@ -12,5 +12,7 @@ public interface StockAPIClient {
 
     Stats getSymbolStats(Symbol symbol) throws ClientException;
 
-    List<HistoricalPrice> getSymbolHistoricalPricesLastFiveYears(Symbol symbol) throws ClientException;
+    List<Price> getSymbolPricesLastFiveYears(Symbol symbol) throws ClientException;
+
+    List<Price> getMarketPricesPreviousDay() throws ClientException;
 }
