+++ src/main/java/com/jespinel/stockreader/clients/iex/IEXClient.java
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.base.Strings;
 import com.jespinel.stockreader.clients.ClientException;
 import com.jespinel.stockreader.clients.StockAPIClient;
-import com.jespinel.stockreader.entities.HistoricalPrice;
+import com.jespinel.stockreader.entities.Price;
 import com.jespinel.stockreader.entities.Stats;
 import com.jespinel.stockreader.entities.Symbol;
 import org.slf4j.Logger;
@@ -62,7 +62,7 @@ public class IEXClient implements StockAPIClient {
     @Override
     public List<Symbol> getSymbols() throws ClientException {
         try {
-            String url = "%s/ref-data/symbols?token=%s".formatted(host, apiKey);
+            String url = "%s/ref-data/iex/symbols?token=%s".formatted(host, apiKey);
             HttpRequest request = HttpRequest.newBuilder()
                     .uri(new URI(url))
                     .timeout(Duration.of(30, SECONDS))
@@ -117,7 +117,7 @@ public class IEXClient implements StockAPIClient {
     }
 
     @Override
-    public List<HistoricalPrice> getSymbolHistoricalPricesLastFiveYears(Symbol symbol) throws ClientException {
+    public List<Price> getSymbolPricesLastFiveYears(Symbol symbol) throws ClientException {
         try {
             String timeRange = "5y";
             String url = "%s/stock/%s/chart/%s?token=%s"
@@ -131,15 +131,44 @@ public class IEXClient implements StockAPIClient {
             HttpClient client = HttpClient.newHttpClient();
             HttpResponse<String> response = client.send(request, HttpResponse.BodyHandlers.ofString());
             if (response.statusCode() != HttpStatus.OK.value()) {
-                String errorMessage = "Error %s getting %s historical prices from IEX Cloud. Details: url: %s, response: %s"
+                String errorMessage = "Error %s getting %s prices from IEX Cloud. Details: url: %s, response: %s"
                         .formatted(response.statusCode(), symbol.getSymbol(), url, response.body());
                 log.error(errorMessage);
                 throw new ClientException(errorMessage);
             }
 
-            TypeReference<List<IEXHistoricalPrice>> symbolList = new TypeReference<>() {};
-            List<IEXHistoricalPrice> iexPrices = objectMapper.readValue(response.body(), symbolList);
-            return iexPrices.stream().map(IEXHistoricalPrice::toHistoricalPrice).collect(Collectors.toList());
+            TypeReference<List<IEXPrice>> symbolList = new TypeReference<>() {};
+            List<IEXPrice> iexPrices = objectMapper.readValue(response.body(), symbolList);
+            return iexPrices.stream().map(IEXPrice::toPrice).collect(Collectors.toList());
+
+        } catch (IOException | InterruptedException | URISyntaxException e) {
+            log.error(e.getMessage(), e);
+            throw new ClientException(e.getMessage(), e);
+        }
+    }
+
+    @Override
+    public List<Price> getMarketPricesPreviousDay() throws ClientException {
+        try {
+            String url = "%s/stock/market/previous?token=%s".formatted(host, apiKey);
+            HttpRequest request = HttpRequest.newBuilder()
+                    .uri(new URI(url))
+                    .timeout(Duration.of(30, SECONDS))
+                    .GET()
+                    .build();
+
+            HttpClient client = HttpClient.newHttpClient();
+            HttpResponse<String> response = client.send(request, HttpResponse.BodyHandlers.ofString());
+            if (response.statusCode() != HttpStatus.OK.value()) {
+                String errorMessage = "Error %s getting market prices from IEX Cloud. Details: url: %s, response: %s"
+                        .formatted(response.statusCode(), url, response.body());
+                log.error(errorMessage);
+                throw new ClientException(errorMessage);
+            }
+
+            TypeReference<List<IEXPrice>> symbolList = new TypeReference<>() {};
+            List<IEXPrice> iexPrices = objectMapper.readValue(response.body(), symbolList);
+            return iexPrices.stream().map(IEXPrice::toPrice).collect(Collectors.toList());
 
         } catch (IOException | InterruptedException | URISyntaxException e) {
             log.error(e.getMessage(), e);
