+++ src/test/resources/application.properties
@@ -13,4 +13,7 @@ stockApiProvider=iex
 ## IEX Cloud
 iex.api_key=test_api_key_123
 iex.host=http://localhost:18080
+# MockServer
+mockserver.logLevel=WARN
+
