--- src/test/resources/mockserver_responses/get_symbol_stats.json
@@ -1,34 +1,0 @@
-{
-  "companyName": "Apple Inc",
-  "marketcap": 2872053733394,
-  "week52high": 186.79,
-  "week52low": 120.64,
-  "week52highSplitAdjustOnly": 190.14,
-  "week52lowSplitAdjustOnly": 118.53,
-  "week52change": 0.3747900306567514,
-  "sharesOutstanding": 16846238295,
-  "float": 0,
-  "avg10Volume": 88316954,
-  "avg30Volume": 84673920,
-  "day200MovingAvg": 160.1,
-  "day50MovingAvg": 176.63,
-  "employees": 147073,
-  "ttmEPS": 11.67,
-  "ttmDividendRate": 0.8848071071918425,
-  "dividendYield": 0.005107671791083752,
-  "nextDividendDate": "",
-  "exDividendDate": "2021-10-23",
-  "nextEarningsDate": "2022-01-21",
-  "peRatio": 15.265667005447652,
-  "beta": 1.33053134418091,
-  "maxChangePercent": 67.64035494925021,
-  "year5ChangePercent": 5.151707365764323,
-  "year2ChangePercent": 1.260454313105051,
-  "year1ChangePercent": 0.3709447327411298,
-  "ytdChangePercent": -0.026111308293420097,
-  "month6ChangePercent": 0.19245381602885225,
-  "month3ChangePercent": 0.2030165158095384,
-  "month1ChangePercent": 0.011326357829007575,
-  "day30ChangePercent": 0.01173997334461943,
-  "day5ChangePercent": 0.005297837368681359
-}
