--- src/test/resources/mockserver_responses/get_symbols.json
@@ -1,56 +1,0 @@
-[
-  {
-    "symbol": "A",
-    "exchange": "XNYS",
-    "exchangeSuffix": "",
-    "exchangeName": "xIoctw Nreoan YcnEkkgc  eSh",
-    "exchangeSegment": "YXNS",
-    "exchangeSegmentName": "x cI cekhkgtnonar cSeEwo NY",
-    "name": "e nintihlnceoo.Iels cTAgg",
-    "date": "2021-12-25",
-    "type": "cs",
-    "iexId": "IEX_46574843354B2D52",
-    "region": "US",
-    "currency": "USD",
-    "isEnabled": true,
-    "figi": "GV320BCDB006",
-    "cik": "1096180",
-    "lei": "8PGA0XWMI2X9WR2Q77UY"
-  },
-  {
-    "symbol": "AA",
-    "exchange": "XNYS",
-    "exchangeSuffix": "",
-    "exchangeName": "ItoccoSagewrkk nEeNxYh cn  ",
-    "exchangeSegment": "SXYN",
-    "exchangeSegmentName": "kInnNt  EwxccraoYkohgeS c e",
-    "name": "Cr olaAocp",
-    "date": "2021-12-25",
-    "type": "cs",
-    "iexId": "IEX_4238333734532D52",
-    "region": "US",
-    "currency": "USD",
-    "isEnabled": true,
-    "figi": "3HBBTD0303BG",
-    "cik": "1743093",
-    "lei": "59W2ZP6930F102T1WUE4"
-  },
-  {
-    "symbol": "AAA",
-    "exchange": "ARCX",
-    "exchangeSuffix": "",
-    "exchangeName": "ascAN eyr",
-    "exchangeSegment": "ACXR",
-    "exchangeSegmentName": "Nsca erAy",
-    "name": "tF oLd-tnEFsirLFdsBT riuyrF  Cd  AioPrTenutit sOs  A",
-    "date": "2021-12-25",
-    "type": "et",
-    "iexId": "IEX_5030314338392D52",
-    "region": "US",
-    "currency": "USD",
-    "isEnabled": true,
-    "figi": "BBX05GF0P48S",
-    "cik": "1725966",
-    "lei": null
-  }
-]
