+++ src/test/resources/mockserver_responses/symbol_prices.json
@@ -1,0 +1,142 @@
+[
+  {
+    "close": 32.5175,
+    "high": 33.6228,
+    "low": 31.8305,
+    "open": 31.8305,
+    "symbol": "AAPL",
+    "volume": 450427610,
+    "id": "LIHSATORISPECIRC_",
+    "key": "ALPA",
+    "subkey": "",
+    "date": "2017-02-01",
+    "updated": 1660153920616,
+    "changeOverTime": 0,
+    "marketChangeOverTime": 0,
+    "uOpen": 129.07,
+    "uClose": 132.8,
+    "uHigh": 132.2,
+    "uLow": 128.42,
+    "uVolume": 117206641,
+    "fOpen": 30.0731,
+    "fClose": 30.6431,
+    "fHigh": 31.232,
+    "fLow": 29.972,
+    "fVolume": 465511602,
+    "label": "Feb 1, 17",
+    "change": 0,
+    "changePercent": 0
+  },
+  {
+    "close": 33.4864,
+    "high": 33.4864,
+    "low": 32.964,
+    "open": 33.4073,
+    "symbol": "AAPL",
+    "volume": 136013846,
+    "id": "ILRC_HISTPESCIROA",
+    "key": "PALA",
+    "subkey": "",
+    "date": "2017-02-02",
+    "updated": 1670984047547,
+    "changeOverTime": -0.001728636236891704,
+    "marketChangeOverTime": -0.001778208225018087,
+    "uOpen": 130.513,
+    "uClose": 128.91,
+    "uHigh": 134.65,
+    "uLow": 129.02,
+    "uVolume": 34129396,
+    "fOpen": 30.3272,
+    "fClose": 30.9123,
+    "fHigh": 31.951,
+    "fLow": 31.3146,
+    "fVolume": 139964119,
+    "label": "Feb 2, 17",
+    "change": -0.05655701645550849,
+    "changePercent": -0.0017
+  },
+  {
+    "close": 32.55,
+    "high": 32.96,
+    "low": 32.55,
+    "open": 32.8421,
+    "symbol": "AAPL",
+    "volume": 100147249,
+    "id": "LEOIACSRTS_CIRPIH",
+    "key": "PAAL",
+    "subkey": "",
+    "date": "2017-02-03",
+    "updated": 1656937360489,
+    "changeOverTime": 0.002640351274290571,
+    "marketChangeOverTime": 0.002583870656843127,
+    "uOpen": 129.63,
+    "uClose": 129.88,
+    "uHigh": 130.38,
+    "uLow": 132.44,
+    "uVolume": 25695995,
+    "fOpen": 30.886,
+    "fClose": 30.7221,
+    "fHigh": 31.893,
+    "fLow": 31.3534,
+    "fVolume": 100634796,
+    "label": "Feb 3, 17",
+    "change": 0.1402433000195536,
+    "changePercent": 0.0044
+  },
+  {
+    "close": 32.9272,
+    "high": 33.0337,
+    "low": 32.686,
+    "open": 33.0337,
+    "symbol": "AAPL",
+    "volume": 108365969,
+    "id": "IACICSHTRPSO_LREI",
+    "key": "APAL",
+    "subkey": "",
+    "date": "2017-02-06",
+    "updated": 1647161185627,
+    "changeOverTime": 0.012266431718437101,
+    "marketChangeOverTime": 0.01202480927186972,
+    "uOpen": 131.38,
+    "uClose": 135.52,
+    "uHigh": 131.8,
+    "uLow": 130.2,
+    "uVolume": 27055270,
+    "fOpen": 31.6127,
+    "fClose": 30.8576,
+    "fHigh": 32.1601,
+    "fLow": 31.3573,
+    "fVolume": 111067666,
+    "label": "Feb 6, 17",
+    "change": 0.314558584898553,
+    "changePercent": 0.0097
+  },
+  {
+    "close": 33.0521,
+    "high": 34.0866,
+    "low": 32.891,
+    "open": 32.891,
+    "symbol": "AAPL",
+    "volume": 153099036,
+    "id": "TRLCHE_AOIIIRCSSP",
+    "key": "PAAL",
+    "subkey": "",
+    "date": "2017-02-07",
+    "updated": 1664061128968,
+    "changeOverTime": 0.022323154516914672,
+    "marketChangeOverTime": 0.02257190147237108,
+    "uOpen": 131.26,
+    "uClose": 137.65,
+    "uHigh": 132.82,
+    "uLow": 134.51,
+    "uVolume": 38561049,
+    "fOpen": 31.1219,
+    "fClose": 32.4307,
+    "fHigh": 32.028,
+    "fLow": 31.16,
+    "fVolume": 154373826,
+    "label": "Feb 7, 17",
+    "change": 0.3141922932032197,
+    "changePercent": 0.0096
+  }
+]
